Renamed provider modules
